Revert "DL-77 : Udpate idea card to inlclude awards"
This reverts commit 761e7e382275187d25ece44aa026c58358f42eb3.

+++ src/app/domain/idea/components/idea-card/idea-card.component.ts
@@ -1,33 +1,33 @@
 /**
  * Created by AKuzmanoski on 26/10/2016.
  */
-import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-import {Idea, Problem} from '../../../model/ideas';
-import {User} from '../../../model/authentication';
-import {IdeaActionsService} from '../../../services/idea/idea-actions.service';
-import {Actions} from '../../../../core/helper/actions/actions';
-import {Action} from '../../../../core/helper/actions/action';
-import {RedirectService} from '../../../../core/navigation/redirect.service';
-
+import {Component, Input, Output, EventEmitter, OnInit} from "@angular/core";
+import {Idea} from "../../../model/ideas/idea";
+import {User} from "../../../model/authentication/user";
+import {Problem} from "../../../model/ideas/problem";
+import {IdeaActionsService} from "../../../services/idea/idea-actions.service";
+import {Actions} from "../../../../core/helper/actions/actions";
+import {Action} from "../../../../core/helper/actions/action";
+import {RedirectService} from "../../../../core/navigation/redirect.service";
 @Component({
   moduleId: module.id,
-  selector: 'ideal-idea-card',
-  templateUrl: 'idea-card.component.html',
-  styleUrls: ['idea-card.component.scss']
+  selector: "ideal-idea-card",
+  templateUrl: "idea-card.component.html",
+  styleUrls: ["idea-card.component.scss"]
 })
 export class IdeaCardComponent implements OnInit {
-  @Input('idea') idea: Idea;
-  @Output('openContent') openContent: EventEmitter<void> = new EventEmitter<void>();
-  @Output('ideaSelected') ideaSelected: EventEmitter<Idea> = new EventEmitter<Idea>();
-  @Output('ideaOwnerSelected') ideaOwnerSelected: EventEmitter<User> = new EventEmitter<User>();
-  @Output('ideaProblemSelected') ideaProblemSelected: EventEmitter<Problem> = new EventEmitter<Problem>();
-  @Output('announce') announce: EventEmitter<Idea> = new EventEmitter<Idea>();
-  @Output('sendTo') sendTo: EventEmitter<Idea> = new EventEmitter<Idea>();
-  @Output('share') share: EventEmitter<Idea> = new EventEmitter<Idea>();
-  @Output('edit') edit: EventEmitter<Idea> = new EventEmitter<Idea>();
-  @Output('report') report: EventEmitter<Idea> = new EventEmitter<Idea>();
-  @Output('remove') remove: EventEmitter<Idea> = new EventEmitter<Idea>();
-  @Output('ban') ban: EventEmitter<Idea> = new EventEmitter<Idea>();
+  @Input("idea") idea: Idea;
+  @Output("openContent") openContent: EventEmitter<void> = new EventEmitter<void>();
+  @Output("ideaSelected") ideaSelected: EventEmitter<Idea> = new EventEmitter<Idea>();
+  @Output("ideaOwnerSelected") ideaOwnerSelected: EventEmitter<User> = new EventEmitter<User>();
+  @Output("ideaProblemSelected") ideaProblemSelected: EventEmitter<Problem> = new EventEmitter<Problem>();
+  @Output("announce") announce: EventEmitter<Idea> = new EventEmitter<Idea>();
+  @Output("sendTo") sendTo: EventEmitter<Idea> = new EventEmitter<Idea>();
+  @Output("share") share: EventEmitter<Idea> = new EventEmitter<Idea>();
+  @Output("edit") edit: EventEmitter<Idea> = new EventEmitter<Idea>();
+  @Output("report") report: EventEmitter<Idea> = new EventEmitter<Idea>();
+  @Output("remove") remove: EventEmitter<Idea> = new EventEmitter<Idea>();
+  @Output("ban") ban: EventEmitter<Idea> = new EventEmitter<Idea>();
   actions: Actions;
 
   constructor(private ideaActionsService: IdeaActionsService, private route: RedirectService) {
@@ -95,48 +95,31 @@ export class IdeaCardComponent implements OnInit {
   }
 
   onAction(action: Action) {
-    switch (action.title) {
-      case 'Details':
-        this.getDetails();
-        break;
-      case 'Problem':
-        this.getProblem();
-        break;
-      case 'Owner':
-        this.getOwner();
-        break;
-      case 'Upvote':
-        this.onUpvote();
-        break;
-      case 'Downvote':
-        this.onDownvote();
-        break;
-      case 'Announce':
-        this.onAnnounce();
-        break;
-      case 'Send':
-        this.onSendTo();
-        break;
-      case 'Share':
-        this.onShare();
-        break;
-      case 'Edit':
-        this.onEdit();
-        break;
-      case 'Remove':
-        this.onRemove();
-        break;
-      case 'Report':
-        this.onReport();
-        break;
-      case 'Ban':
-        this.onBan();
-        break;
-      case 'Buy':
-        this.onBuy();
-        break;
-      default:
-        break;
-    }
+    if (action.title == "Details")
+      this.getDetails();
+    else if (action.title == "Problem")
+      this.getProblem();
+    else if (action.title == "Owner")
+      this.getOwner();
+    else if (action.title == "Upvote")
+      this.onUpvote();
+    else if (action.title == "Downvote")
+      this.onDownvote();
+    else if (action.title == "Announce")
+      this.onAnnounce();
+    else if (action.title == "Send")
+      this.onSendTo();
+    else if (action.title == "Share")
+      this.onShare();
+    else if (action.title == "Edit")
+      this.onEdit();
+    else if (action.title == "Remove")
+      this.onRemove();
+    else if (action.title == "Report")
+      this.onReport();
+    else if (action.title == "Ban")
+      this.onBan();
+    else if (action.title == "Buy")
+      this.onBuy();
   }
 }
